[feat] create real time plot widget
